+++ tests/opencertserver.est.server.tests/README.md
@@ -1,0 +1,3 @@
+# Test Steps for EST Server
+
+This file contains test steps for the EST server using ReqnRoll.
