+++ tests/opencertserver.lambda.tests/README.md
@@ -1,0 +1,3 @@
+# Test Steps for Lambda Function
+
+This file contains test steps for the Lambda function using ReqnRoll.
